# backend/openapi.json
{
  "components": {
    "schemas": {
      "Action": {
        "description": "What the agent did, or what ground truth says it should have done.\n\nESCALATE is only ever an *agent* action. Ground truth is always APPROVE or REJECT,\nbecause every synthetic invoice carries a deterministic correct answer.",
        "enum": [
          "APPROVE",
          "REJECT",
          "ESCALATE"
        ],
        "title": "Action",
        "type": "string"
      },
...
      "DecisionCreate": {
        "description": "Request body for `POST /api/v1/decisions` \u2014 the simulator's ingest path.\n\nEvery field needed to construct a `DecisionRecord` (see\n`shared.contracts.DecisionRecord`), plus `reason`\n(cross-cutting rule: every state-changing endpoint requires one \u2014 here,\nwhy this decision is being submitted, e.g. which simulation run or\ninvoice-processing batch produced it).",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "amount": {
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "type": "integer"
          },
          "ground_truth": {
            "$ref": "#/components/schemas/Action"
          },
          "invoice_id": {
            "title": "Invoice Id",
            "type": "string"
          },
          "reason": {
            "description": "Why this decision is being submitted",
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "invoice_id",
          "amount",
          "action",
          "ground_truth",
          "agent_id",
          "reason"
        ],
        "title": "DecisionCreate",
        "type": "object"
      },
      "DecisionRecordOut": {
        "description": "Mirrors `shared.contracts.DecisionRecord` field-for-field.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "amount": {
            "title": "Amount",
            "type": "integer"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "ground_truth": {
            "$ref": "#/components/schemas/Action"
          },
          "human_ruling": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Action"
              },
              {
                "type": "null"
              }
            ]
          },
          "invoice_id": {
            "title": "Invoice Id",
            "type": "string"
          },
          "recommended_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Action"
              },
              {
                "type": "null"
              }
            ]
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "decision_id",
          "sequence",
          "invoice_id",
          "amount",
          "action",
          "ground_truth",
          "agent_id",
          "decided_at",
          "recommended_action",
          "human_ruling"
        ],
        "title": "DecisionRecordOut",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Every non-2xx response returns this shape as its JSON body.\n\n`code` is a short machine-readable slug (e.g. \"not_found\", \"forbidden\",\n\"validation_error\") for client branching; `message` is one sentence a human\nreads; `detail` carries whatever extra structure the specific failure needs\n(field errors, the resource id that was missing, etc.) and may be null.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_DecisionRecordOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DecisionRecordOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "1-indexed page number",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Items per page, as requested",
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matching records, independent of page size",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[DecisionRecordOut]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The integration point for all four lanes: the Policy Engine, persistence and audit, the HTTP contract, and the hard ceiling. See docs/SYSTEM-EXPLAINED.md for the full design.",
    "title": "Adaptive AI Governance Platform \u2014 Backend",
    "version": "1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/decisions": {
      "get": {
        "description": "List decisions, newest first.\n\n`?agent_id=` filters in SQL. The agent detail page used to fetch the\nnewest 50 decisions across every agent and filter them in the browser,\nwhich silently showed nothing at all once another agent's run pushed it\noff the first page \u2014 a blank panel that looked like \"no decisions\" rather\nthan \"wrong query\".",
        "operationId": "list_decisions_api_v1_decisions_get",
        "parameters": [
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DecisionRecordOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Decisions",
        "tags": [
          "decisions"
        ]
      },
      "post": {
        "description": "Ingest one decision from the simulator (or, eventually, a real agent).\n\nValidates the agent exists and has a policy version on record, evaluates\nthe invoice against that policy with the real Policy Engine\n(`app.policy.engine.evaluate_decision`), and \u2014 in one transaction \u2014\npersists the invoice (if new), the decision (referencing the exact\npolicy version in force), and a hash-chained `audit_log` entry\n(docs/lanes/vp.md). `within_limit` is recorded on the decision row\nverbatim from the Policy Engine's own output; the Policy Engine's reason\ncode is recorded on the audit entry, never recomputed later.",
        "operationId": "create_decision_api_v1_decisions_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionRecordOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The current role may not perform this action."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Decision",
        "tags": [
          "decisions"
        ]
      }
    },
    "/api/v1/decisions/{decision_id}": {
      "get": {
        "description": "Fetch one decision by id.",
        "operationId": "get_decision_api_v1_decisions__decision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionRecordOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Decision",
        "tags": [
          "decisions"
        ]
      }
    },
...
  }
}